requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-on-heroku==1.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
whitenoise==6.3.0
